+++ graphdatascience/tests/unit/test_error_handling.py
@@ -15,7 +15,7 @@ from graphdatascience.tests.unit.resources.example_server_endpoints import (
 )
 
 
-@pytest.mark.parametrize("server_version", [ServerVersion(2, 2, 0)])  # Something later than 2.1.0
+@pytest.mark.parametrize("server_version", [ServerVersion(2, 10, 0)])  # Something later than 2.1.0
 def test_incompatible_server_version(runner: QueryRunner, gds: GraphDataScience) -> None:
     G = Graph("dummy", runner)
     with pytest.raises(